Polishing RayTracing
