+++ ChiTech/ChiMesh/LogicalVolume/chi_mesh_logicalvolume_surfmesh.cc
@@ -3,7 +3,7 @@
 #include "chi_mesh_logicalvolume.h"
 #include "ChiMesh/chi_mesh.h"
 #include "ChiMesh/SurfaceMesh/chi_surfacemesh.h"
-#include "ChiMesh/Raytrace/raytracing_old.h"
+#include "ChiMesh/Raytrace/raytracing.h"
 
 //###################################################################
 /**Constructor to compute bound box information.*/
