+++ ChiTech/ChiMesh/Raytrace/raytrace_utils.cc
@@ -1,4 +1,4 @@
-#include "raytracing_old.h"
+#include "raytracing.h"
 #include "ChiMesh/Cell/cell.h"
 #include "ChiMesh/MeshContinuum/chi_meshcontinuum.h"
