+++ ChiTech/ChiMesh/Raytrace/raytracer.cc
@@ -1,4 +1,4 @@
-#include "raytracer.h"
+#include "raytracing.h"
 
 #include "ChiMesh/MeshContinuum/chi_meshcontinuum.h"
